Update Person.swift
I changed the creatOrder() function in Order.swift to allow items with zero quantities to be added to the order so that even those will be checked when determining if the order is correct or not.  I also changed the setTextForToolTip() function in Person.swift to replace the text for zero quantity items in the tooltip.  “0 ketchups” becomes “no ketchups” and so on.  I changed  touchesBegan() in GameScene.swift to prevent a bug that crashes when you click on a person.